Separating GameArtworkService with GameCover- and GameIconService. I am also updated the work for the Views with new sizes and updated the ModView to fit the covers. I also implemented WinHttpHandler nuget package. Still working on finishing touches.
