+++ src/Restall.Application/Interfaces/Driven/IGameCoverService.cs
@@ -1,0 +1,8 @@
+using Restall.Domain.Entities;
+
+namespace Restall.Application.Interfaces.Driven;
+
+public interface IGameCoverService
+{
+    Task DownloadCoverIfMissingAsync(Game game, string coverPath);
+}
