+++ src/Restall.Application/Interfaces/Driven/IGameIconService.cs
@@ -1,0 +1,6 @@
+namespace Restall.Application.Interfaces.Driven;
+
+public interface IGameIconService
+{
+    Task ExtractIconIfMissingAsync(string? executablePath, string? gameName, string iconPath);
+}
