+++ src/Restall.UI/Views/ModView.axaml
@@ -3,82 +3,64 @@
              xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
              xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
              xmlns:vm="clr-namespace:Restall.UI.ViewModels"
-             mc:Ignorable="d" d:DesignWidth="1000" d:DesignHeight="450"
+             mc:Ignorable="d" d:DesignWidth="800" d:DesignHeight="450"
              x:Class="Restall.UI.Views.ModView"
              x:DataType="vm:ModViewModel">
 
     <Grid ColumnDefinitions="*"
-          RowDefinitions=".45*,.55*">
+          RowDefinitions="Auto,*">
 
         <!-- GAME CARD -->
-        <Border Grid.Row="0" CornerRadius="20" Margin="5,5,5,5" Background="{DynamicResource GameCardBackground}">
-
+        <Border Grid.Row="0" CornerRadius="20" Margin="5" Background="{DynamicResource GameCardBackground}">
             <Grid ColumnDefinitions="Auto,*">
-
-                <!-- Cover Art -->
-                <Border Grid.Column="0" CornerRadius="14" Width="350">
+                
+                <!--- GAME COVER -->
+                <Border Grid.Column="0" CornerRadius="6" Margin="10" ClipToBounds="True"
+                        Width="200" Height="280"
+                        VerticalAlignment="Top">
                     <Image Source="{Binding SelectedGame.CoverBitmap}"
-                           Stretch="UniformToFill" MinWidth="400"
-                           VerticalAlignment="Stretch" />
+                           Stretch="UniformToFill"
+                           VerticalAlignment="Center"
+                           HorizontalAlignment="Center" />
                 </Border>
-
-                <!-- Game Info -->
+                
+                <!--- GAME INFO -->
                 <Grid Grid.Column="1"
-                      ColumnDefinitions="*,Auto"
-                      RowDefinitions="Auto,Auto,Auto,Auto,Auto,Auto"
-                      Margin="10,10,10,0">
-
-                    <StackPanel Grid.Column="0" Grid.Row="0" Grid.ColumnSpan="2"
-                                Orientation="Horizontal" Spacing="6" VerticalAlignment="Center">
-                        <TextBlock Text="{Binding SelectedGame.Name}" FontSize="20" FontWeight="Bold" Margin="20,0,0,0"
-                                   Foreground="{DynamicResource GameCardText}" TextWrapping="Wrap"
-                                   VerticalAlignment="Center" />
-                        <Button Classes="game-card-btn" CornerRadius="6"
-                                Command="{Binding OpenInExplorerCommand}"
-                                VerticalAlignment="Center">
-                            <TextBlock FontFamily="{StaticResource Phosphor-Fill}" FontSize="20" Text="&#xE256;" />
-                        </Button>
-                    </StackPanel>
-
-                    <TextBlock Grid.Column="0" Grid.ColumnSpan="2" Grid.Row="1" FontSize="10" Margin="20,0,0,0"
-                               Text="{Binding SelectedGame.PlatformName}" Foreground="{DynamicResource GameCardText}"
+                      ColumnDefinitions="*"
+                      RowDefinitions="Auto,Auto,Auto,Auto, Auto"
+                      Margin="8"
+                      VerticalAlignment="Top">
+        
+                    <TextBlock Grid.Row="0" Grid.ColumnSpan="2" Text="{Binding SelectedGame.Name}"
+                               FontSize="20" FontWeight="Bold"
+                               Foreground="{DynamicResource GameCardText}"
                                TextWrapping="Wrap" />
-                    <TextBlock Grid.Column="0" Grid.ColumnSpan="2" Grid.Row="2" FontSize="10" Margin="20,0,0,0"
-                               Text="{Binding SelectedGame.EngineName}" Foreground="{DynamicResource GameCardText}"
+        
+                    <TextBlock Grid.Row="1" Text="{Binding SelectedGame.PlatformName}"
+                               FontSize="12" Foreground="{DynamicResource GameCardText}" />
+        
+                    <TextBlock Grid.Row="2" Text="{Binding SelectedGame.EngineName}"
+                               FontSize="12" Foreground="{DynamicResource GameCardText}" />
+        
+                    <TextBlock Grid.Row="3" Text="{Binding SelectedGame.InstallFolderDisplay}"
+                               FontSize="12" Foreground="{DynamicResource GameCardText}"
                                TextWrapping="Wrap" />
-                    <TextBlock Grid.Column="0" Grid.ColumnSpan="2" Grid.Row="3" FontSize="10" Margin="20,0,0,0"
-                               Text="{Binding SelectedGame.InstallFolder}" Foreground="{DynamicResource GameCardText}"
-                               TextWrapping="Wrap" />
-                    <TextBlock Grid.Column="0" Grid.ColumnSpan="2" Grid.Row="4" FontSize="10" Margin="20,0,0,0"
-                               Text="{Binding SelectedGame.ExecutablePath}" Foreground="{DynamicResource GameCardText}"
-                               TextWrapping="Wrap" />
-
-                    <Grid Grid.Column="0" Grid.Row="5" Grid.ColumnSpan="2"
-                          ColumnDefinitions="*,*,*" ColumnSpacing="5"
-                          Margin="5,30,5,5">
-                        <Button Grid.Column="0" Classes="game-card-btn" CornerRadius="6"
-                                VerticalAlignment="Center" HorizontalAlignment="Stretch"
-                                Margin="0,0,4,0" IsVisible="False">
-                            <TextBlock Text="Quick game scan" FontSize="14" HorizontalAlignment="Center" />
+        
+                    <Border Grid.Row="4" CornerRadius="6" Background="{DynamicResource GameCardNotesBorder}"
+                            Padding="4,4" Margin="0,6,0,0">
+                        <Button Classes="game-card-btn" Command="{Binding OpenInExplorerCommand}"
+                                Padding="0" Background="Transparent" BorderThickness="0">
+                            
+                                <TextBlock
+                                           Text="{Binding SelectedGame.ExecutablePathDisplay}"
+                                           FontSize="12" Foreground="{DynamicResource GameCardText}"
+                                           Opacity="0.8" FontStyle="Italic"
+                                           TextWrapping="Wrap"/>
                         </Button>
-                        <Button Grid.Column="1" Classes="game-card-btn" CornerRadius="6"
-                                VerticalAlignment="Center" HorizontalAlignment="Stretch"
-                                HorizontalContentAlignment="Center" Margin="0,0,4,0" IsVisible="False">
-                            <TextBlock Text="Display Commander" FontSize="14" HorizontalAlignment="Center" />
-                        </Button>
-                        <Button Grid.Column="2" Classes="game-card-btn" CornerRadius="6"
-                                VerticalAlignment="Center" HorizontalAlignment="Stretch"
-                                HorizontalContentAlignment="Center" IsVisible="False">
-                            <TextBlock Text="SpecialK" FontSize="14" HorizontalAlignment="Center" />
-                        </Button>
-                    </Grid>
-
+                    </Border>
                 </Grid>
             </Grid>
-
-
         </Border>
-
         <!-- MOD CARD -->
         <Border Grid.Row="1" Margin="5" CornerRadius="20" Background="{DynamicResource GameCardBackground}">
             <Grid RowDefinitions="Auto, Auto"
